chore: merge shared settings into polyfills
