+++ app/renderer/polyfills/index.js
@@ -1,3 +1,21 @@
+export const PRESETS = 'presets';
+export const SAVED_SESSIONS = 'SAVED_SESSIONS';
+export const SET_SAVED_GESTURES = 'SET_SAVED_GESTURES';
+export const SERVER_ARGS = 'SERVER_ARGS';
+export const SESSION_SERVER_PARAMS = 'SESSION_SERVER_PARAMS';
+export const SESSION_SERVER_TYPE = 'SESSION_SERVER_TYPE';
+export const SAVED_FRAMEWORK = 'SAVED_FRAMEWORK';
+
+export const DEFAULT_SETTINGS = {
+  [PRESETS]: {},
+  [SAVED_SESSIONS]: [],
+  [SET_SAVED_GESTURES]: [],
+  [SERVER_ARGS]: null,
+  [SESSION_SERVER_PARAMS]: null,
+  [SESSION_SERVER_TYPE]: null,
+  [SAVED_FRAMEWORK]: 'java',
+};
+
 let log, settings, clipboard, shell, remote, ipcRenderer,
     i18NextBackend, i18NextBackendOptions, fs, util;
 
@@ -21,6 +39,17 @@ if (buildForBrowser()) {
   ({log, settings, clipboard, shell, remote,
     ipcRenderer, i18NextBackend,
     i18NextBackendOptions, fs, util} = require('./electron'));
+}
+
+export async function getSetting (setting) {
+  if (await settings.has(setting)) {
+    return await settings.get(setting);
+  }
+  return DEFAULT_SETTINGS[setting];
+}
+
+export async function setSetting (setting, value) {
+  await settings.set(setting, value);
 }
 
 export {
