+++ app/renderer/routes.js
@@ -4,9 +4,8 @@ import App from './containers/App';
 import SessionPage from './containers/SessionPage';
 import InspectorPage from './containers/InspectorPage';
 import Spinner from './components/Spinner/Spinner';
-import { ipcRenderer } from './polyfills';
+import { ipcRenderer, setSetting } from './polyfills';
 import i18n from '../configs/i18next.config.renderer';
-import { setSetting } from '../shared/settings';
 
 ipcRenderer.on('appium-language-changed', (_evt, message) => {
   setSetting('PREFERRED_LANGUAGE', message.language);
